src/models/train.py
"""GNN training entry point for DCN fault detection.

Loads a processed PyG graph dataset from data/processed/, instantiates the
model architecture specified in the training config, and runs the training
loop with periodic validation and checkpoint saving.

Design principles:
- Fully config-driven: all hyperparameters come from a YAML training config.
- Reproducible: random seeds are set and logged before any stochastic ops.
- The dataset (data/processed/) is treated as a fixed artefact; this module
  never calls simulation or graph-building code.
- Checkpoints are saved to configs-specified output directories so that
  evaluation (evaluate.py) can load any checkpoint independently.
"""

from __future__ import annotations

import logging
import random
from pathlib import Path
from typing import Any

import numpy as np
import torch
import yaml

logger = logging.getLogger(__name__)


def set_seeds(seed: int) -> None:
    """Set random seeds for Python, NumPy, and PyTorch for reproducibility.

    Args:
        seed: Integer seed value. Logged at INFO level.
    """
    raise NotImplementedError


def load_dataset(processed_dir: Path, split: str = "train") -> Any:
    """Load a processed PyG dataset split from disk.

    Args:
        processed_dir: Directory containing .pt graph dataset files.
        split: One of 'train', 'val', 'test'.

    Returns:
        PyG InMemoryDataset or list of Data objects for the requested split.
    """
    raise NotImplementedError


def train(config_path: Path) -> None:
    """Run the full training pipeline from a YAML config file.

    Reads config_path, loads dataset splits, builds the model architecture,
    runs the training loop, validates after each epoch, and saves checkpoints
    to the output directory specified in the config.

    Args:
        config_path: Path to the YAML training configuration file.
    """
    raise NotImplementedError
