pyproject.toml
[tool.poetry]
name = "dcn-benchmark"
version = "0.1.0"
description = "Synthetic benchmark dataset for District Cooling Networks with GNN-based fault detection"
authors = []
readme = "README.md"
packages = [{ include = "generation", from = "src" },
            { include = "processing", from = "src" },
            { include = "models", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
wntr = "*"
networkx = "*"
pandas = "*"
numpy = "*"
pyyaml = "*"

[tool.poetry.group.ml.dependencies]
torch = "*"
torch-geometric = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
